+++ app/src/main/java/com/flex/domain/model/Settings.kt
@@ -27,5 +27,6 @@ data class Settings(
     val calendarSyncOffice: Boolean = true,
     val calendarSyncHomeOffice: Boolean = true,
     val calendarEventPrefix: String = "FleX",
-    val calendarEventNoAlarm: Boolean = false
+    val calendarEventNoAlarm: Boolean = false,
+    val federalState: FederalState = FederalState.HAMBURG
 )